requirement.txt
alembic==1.13.1
Authlib==0.15.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Deprecated==1.2.14
Flask==2.0.1
Flask-Dance==3.0.0
Flask-Limiter==1.4
Flask-Login==0.5.0
flask-mail==0.10.0
Flask-Migrate==3.1.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==3.7
importlib-metadata==8.0.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.1
pycparser==2.22
pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.0.2
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.7.1
six==1.16.0
SQLAlchemy==1.4.35
typing-extensions==4.12.2
urllib3==2.2.2
URLObject==2.4.3
waitress==3.0.0
Werkzeug==2.0.1
wrapt==1.16.0
zipp==3.19.2
